+++ lupyne/engine/utils.py
@@ -1,10 +1,6 @@
 import abc
-import bisect
 import contextlib
-import heapq
-import itertools
 from collections.abc import Iterable
-from typing import Optional
 import lucene
 from java.lang import Double, Float, Number, Object
 from org.apache.lucene import analysis, util
@@ -20,29 +16,6 @@ class Atomic(metaclass=abc.ABCMeta):
 
 for cls in (str, analysis.TokenStream, lucene.JArray_byte):
     Atomic.register(cls)
-
-
-class SpellChecker(dict):
-    """Correct spellings and suggest words for queries.
-
-    Supply a vocabulary mapping words to (reverse) sort keys, such as document frequencies.
-    """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.words = sorted(self)
-        alphabet = ''.join(set(itertools.chain.from_iterable(self)))
-        self.suffix = alphabet and max(alphabet) * max(map(len, self))
-
-    def complete(self, prefix: str, count: Optional[int] = None) -> list:
-        """Return ordered suggested words for prefix."""
-        start = bisect.bisect_left(self.words, prefix)
-        stop = bisect.bisect_right(self.words, prefix + self.suffix, start)
-        words = self.words[start:stop]
-        if count is not None and count < len(words):
-            return heapq.nlargest(count, words, key=self.__getitem__)
-        words.sort(key=self.__getitem__, reverse=True)
-        return words
 
 
 @contextlib.contextmanager