+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Removed
 * Spatial field
+* Cached custom spellchecker
 
 ## [3.1](https://pypi.org/project/lupyne/3.1/) - 2023-11-22
 ### Changed